# codegen/openapi.yaml
openapi: 3.1.0
info:
  title: RPA Control-Plane API
  version: "1.0"
  description: >
    제어평면 REST API (v1). 운영 콘솔/외부 클라이언트가 호출하는 제어평면 엔드포인트의
    경로·메서드·요청/응답 요지·주요 에러코드를 고정한다. 본 문서는 api-surface.md(Control-Plane
    API Surface v1)를 1:1 변환한 D1 codegen 산출물이다.


    요지 수준 규약: components.schemas는 요지(shape) 수준이며, 상세 타입(IR/Verify/Event payload·필드
    required·examples)은 계약 ts/schema 및 codegen/types를 참조한다. 어휘 체인 abort → cancelled →
    run.cancelled, "조용한 false/unknown 금지", error-catalog 메타(httpStatus/exceptionClass)를 그대로 따른다.


    - 인증: 모든 엔드포인트 인증 필수. tenant_id는 요청 본문/쿼리가 아니라 인증 주체(토큰)에서 도출한다(api-surface §0.1).

    - 에러 본문: 모든 4xx/5xx는 ApiError(ts/error-catalog.ts). HTTP 상태코드는 ERROR_CATALOG[code].httpStatus를 그대로 사용한다(api-surface §0.2).

    - DEAD_LETTER(httpStatus 200)는 API 오류 응답이 아니다 — 상태 통지 전용 코드이므로 ApiError로 반환하지 않는다(api-surface §0.2/§4).
servers:
  - url: /v1
    description: Control-plane v1 base path.
...
security:
  - bearerAuth: []
paths:
...
  /human-tasks:
    get:
      tags: [HumanTasks]
      operationId: listHumanTasks
      description: List human-in-the-loop tasks for queue triage and suspended run recovery.
      summary: 휴먼 태스크 인박스 목록
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/HumanTaskState'
        - name: kind
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/HumanTaskKind'
        - name: assignee
          in: query
          required: false
          schema: { type: string }
        - name: run_id
          in: query
          required: false
          description: suspended run 상세에서 연결된 human task로 정확히 딥링크하기 위한 필터.
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: 커서 페이지(items=HumanTask).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanTaskPage'
        default:
          $ref: '#/components/responses/Error'
  /principals:
    get:
      tags: [HumanTasks]
      operationId: listPrincipals
      summary: 테넌트 담당자 디렉터리(name-picker 소스)
      description: >
        테넌트 주체 디렉터리(principals 테이블). 각 항목은 principal_id(surrogate uuid)·sub(PrincipalId=JWT sub)·
        display_name·email·source(jwt|manual). 표시명은 JWT optional name 클레임 자동 upsert(source=jwt) 또는
        admin 수동 등록(source=manual)으로 채운다. human_tasks.assignee→principals FK는 없어 디렉터리 미등록 sub도
        직접 배정 가능(자유 입력 폴백 유지). RBAC는 principal.read(viewer+; 실 배정은 human_task.assign이 강제).
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: 커서 페이지(items=Principal; 공유 (created_at, id) keyset).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrincipalPage'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [HumanTasks]
      operationId: createPrincipal
      summary: 담당자 디렉터리 수동 등록(admin)
      description: >
        admin이 미로그인 담당자를 사전 등록한다(source=manual). UNIQUE(tenant_id, sub) 위반(이미 등록) → 422.
        RBAC는 principal.manage(admin).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrincipalCreateRequest'
      responses:
        '201':
          description: 등록된 디렉터리 항목.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Principal'
        '403':
          $ref: '#/components/responses/Error'        # AUTHZ_FORBIDDEN
        '422':
          $ref: '#/components/responses/Error'        # IR_SCHEMA_INVALID(중복 sub)
        default:
          $ref: '#/components/responses/Error'
...
  /human-tasks/{human_task_id}:
    parameters:
      - $ref: '#/components/parameters/HumanTaskId'
    get:
      tags: [HumanTasks]
      operationId: getHumanTask
      description: Return one human task with assignment, timeout, and related run metadata.
      summary: 휴먼 태스크 상세
      responses:
        '200':
          description: 태스크 상세(state, kind, assignee, timeout, on_timeout, run 연계). v1은 kind별 payload 본문을 노출하지 않는다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanTask'
        '404':
          $ref: '#/components/responses/Error'        # RESOURCE_NOT_FOUND
        default:
          $ref: '#/components/responses/Error'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT bearer. tenant_id 및 roles(Role[]) 클레임에서 도출(auth-rbac §3). 클라이언트 body/query의
        tenant_id는 불신. 인증 미성립(토큰 누락/서명 무효) 시 UNAUTHENTICATED(401), 인증됐으나 tenant
        누락/모호 또는 역할 권한 부족 시 AUTHZ_FORBIDDEN(403). 두 경우 모두 본문은 ApiError.
  # --- Reusable parameters ---
  parameters:
...
    HumanTaskId:
      name: human_task_id
      in: path
      required: true
      schema: { type: string }
...
    Limit:
      name: limit
      in: query
      required: false
      description: 페이지 크기. 상한 기본값은 운영 정책(Phase 3) — 본 계약은 파라미터 형태만 고정(api-surface §0.5).
      schema:
        type: integer
        minimum: 1
    Cursor:
      name: cursor
      in: query
      required: false
      description: 불투명 커서. 직전 응답 next_cursor 값을 그대로 전달(api-surface §0.5).
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >
        부작용 명령형 POST의 중복 제출 보호(api-surface §0.4). 서버는 (tenant_id, endpoint, key)로
        최초 결과를 보관, 동일 키 재제출 시 부작용 재실행 없이 최초 응답 반환. sink_idempotency_key(데이터평면)와는 다른 계층.
      schema:
        type: string

  # --- Reusable headers ---
...
  responses:
    Error:
      description: >
        오류 응답. 본문은 ApiError(ts/error-catalog.ts). HTTP 상태는 ERROR_CATALOG[code].httpStatus를
        그대로 사용한다. DEAD_LETTER는 오류 응답이 아니므로 본문 code로 반환되지 않는다(api-surface §0.2).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  # --- Schemas (요지 수준; 상세 타입은 계약 ts/schema 및 codegen/types 참조) ---
  schemas:
...
    ApiError:
      type: object
      description: 공통 에러 응답(ts/error-catalog.ts ApiError).
      required: [code, message, correlation_id]
      additionalProperties: false
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: userMessage 또는 상세(민감정보 없음).
        details:
          description: 코드별 상세 페이로드(예 IR_SCHEMA_INVALID의 reason). 구조는 D1 위임(api-surface §9).
        correlation_id:
          type: string
          format: uuid
          description: event-envelope·trace correlation_id와 동일 값(api-surface §0.2).
    ErrorCode:
      type: string
      description: >
        ts/error-catalog.ts ErrorCode(48개). DEAD_LETTER는 enum에 포함되나 상태 통지 전용으로
        ApiError 응답 code로는 반환되지 않는다(api-surface §0.2).
      enum:
        - RUN_NOT_FOUND
        - RESOURCE_NOT_FOUND
        - RUN_ALREADY_TERMINAL
        - RUN_ABORTED
        - SCENARIO_VERSION_CONFLICT
        - POLICY_VERSION_CONFLICT
        - IR_SCHEMA_INVALID
        - IR_EXPRESSION_COMPILE_ERROR
        - IR_EXPRESSION_RUNTIME
        - IR_NO_BRANCH_MATCHED
        - SITE_PROFILE_BLOCKED
        - SITE_CIRCUIT_OPEN
        - SESSION_LOCKED
        - SESSION_GENERATION_CONFLICT
        - SESSION_REGISTRATION_REQUIRED
        - CHALLENGE_UNRESOLVED
        - RATE_BUDGET_EXCEEDED
        - BROWSER_LEASE_EXPIRED
        - BROWSER_CRASH
        - CDP_DISCONNECTED
        - NAVIGATION_TIMEOUT
        - LLM_BUDGET_EXCEEDED
        - LLM_CAPABILITY_MISMATCH
        - LLM_STREAM_TIMEOUT
        - LLM_STREAM_IDLE_TIMEOUT
        - LLM_MALFORMED_OUTPUT
        - LLM_CONTENT_FILTERED
        - LLM_RATE_LIMITED
        - LLM_BACKEND_UNAVAILABLE
        - LLM_CONNECTION_FAILED
        - EXTRACT_SCHEMA_INVALID
        - VERIFY_FAILED
        - EMPTY_RESULT_NO_WITNESS
        - SECRET_ACCESS_DENIED
        - DOMAIN_POLICY_VIOLATION
        - PROMPT_INJECTION_DETECTED
        - ARTIFACT_NOT_REDACTED
        - SHELL_COMMAND_NOT_ALLOWED
        - UNAUTHENTICATED
        - AUTHZ_FORBIDDEN
        - CONNECTOR_PERMISSION_DENIED
        - CONNECTOR_INCOMPATIBLE
        - CONNECTOR_HOOK_FAILED
        - SINK_DELIVERY_FAILED
        - RAW_PERSIST_FAILED
        - CONTROL_PLANE_INTERNAL_ERROR
        - HUMAN_TASK_EXPIRED
        - APPROVAL_ALREADY_DECIDED
        - WORKITEM_CHECKOUT_CONFLICT
        - DEAD_LETTER

    # ===== State enums (ts/state-machine-types.ts — 정확히 일치) =====
...
    HumanTaskState:
      type: string
      enum:
        - open
        - assigned
        - in_progress
        - resolved
        - expired
        - cancelled
        - escalated
    HumanTaskKind:
      type: string
      enum:
        - approval
        - validation
        - exception
        - captcha
        - mfa
...
    HumanTaskPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/HumanTask' }
        next_cursor:
          type: [string, "null"]
    Principal:
      type: object
      required: [principal_id, sub, display_name, source]
      description: >
        테넌트 담당자 디렉터리 항목(principals 테이블). principal_id=surrogate uuid, sub=PrincipalId(JWT sub, 자유형),
        display_name=표시이름, email=optional, source=jwt|manual(쓰기 경로).
      properties:
        principal_id: { type: string }
        sub: { type: string }
        display_name: { type: string }
        email: { type: [string, "null"] }
        source: { type: string, enum: [jwt, manual] }
    PrincipalPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Principal' }
        next_cursor:
          type: [string, "null"]
    PrincipalCreateRequest:
      type: object
      required: [sub, display_name]
      description: admin 수동 등록(source=manual). sub=PrincipalId(JWT sub, 자유형), display_name=표시이름.
      properties:
        sub: { type: string, minLength: 1 }
        display_name: { type: string, minLength: 1 }
        email: { type: [string, "null"] }
      additionalProperties: false
...
    HumanTask:
      type: object
      description: 휴먼 태스크 상세. artifact 본문은 별도 Artifacts API의 redaction/RBAC 경계를 탄다.
      properties:
        human_task_id: { type: string }
        state: { $ref: '#/components/schemas/HumanTaskState' }
        kind: { $ref: '#/components/schemas/HumanTaskKind' }
        assignee: { type: [string, "null"] }
        timeout: { type: [string, "null"] }
        on_timeout:
          type: string
          enum: [fail, escalate]
          description: reserved-handlers @human_task 입력(기본 fail). HumanTaskGuard.onTimeout과 정합.
        run_id: { type: [string, "null"] }
        payload:
          type: object
          additionalProperties: true
        result_schema:
          oneOf:
            - $ref: '#/components/schemas/HumanTaskBusinessFormSchema'
            - type: object
              additionalProperties: true
          description: >
            HumanTask V2 form hint. version=business_form_v1 is the configurable HITL/data-entry contract;
            legacy objects without version remain read-only hints.
        artifact_refs:
          type: array
          items: { type: string }
        result:
          oneOf:
            - $ref: '#/components/schemas/HumanTaskResolution'
            - type: "null"
    HumanTaskResolution:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approve, reject, correct, retry]
        corrections:
          type: object
          additionalProperties: true
          description: >
            For result_schema.version=business_form_v1 and decision=correct, keys must match declared fields
            and values must satisfy required/type/options validation.
        reason: { type: string }
        confidence:
          type: number
          minimum: 0
          maximum: 1
        notes: { type: string }
      additionalProperties: false
    HumanTaskBusinessFormSchema:
      type: object
      required: [version, fields]
      properties:
        version:
          type: string
          enum: [business_form_v1]
        title:
          type: string
        description:
          type: string
        fields:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/HumanTaskBusinessFormField'
      additionalProperties: false
    HumanTaskBusinessFormField:
      type: object
      required: [key, label, type]
      properties:
        key:
          type: string
          pattern: '^[A-Za-z_][A-Za-z0-9_]{0,63}$'
        label:
          type: string
        type:
          type: string
          enum: [text, textarea, number, boolean, date, select]
        required:
          type: boolean
        options:
          type: array
          items:
            type: string
        help_text:
          type: string
      additionalProperties: false
...